1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t k;
+	unsigned int k;
 
 	for (k = 0; k < size; k++)
 		action(array[k]);
